Refined bitfield labels
